make the answer value accessed by the extended classes (try 2)
It looks like the answer variable from the base class is not being
updated by the inherited classes, and so there is no value being set
and so the system doesn't record a value for any of the sections (the
values are all undefined).

It seems like that answer value needs to be accessed by the extended
classes.